+++ lib/src/log_stack.dart
@@ -1,19 +1,87 @@
 part of 'logger.dart';
 
 /// AWS Hard Limits
-const int awsMaxByteMessageSize = 262118;
-const int awsMaxByteBatchSize = 1048550;
-const int awsMaxMessagePerByte = 10000;
+const int awsMaxBytesPerMessage = 262116;
+const int awsMinBytesPerMessage = 5;
+const int awsMaxBytesPerRequest = 1048576;
+const int awsMinBytesPerRequest = 1;
+const int awsMaxMessagesPerRequest = 10000;
+const int awsMinMessagesPerRequest = 1;
 
 /// A class that automatically splits and handles logs according to AWS hard limits
 class CloudWatchLogStack {
   /// An enum value that indicates how messages larger than the max size should be treated
   CloudWatchLargeMessages largeMessageBehavior;
 
+  /// Changes how large each message can be before [largeMessageBehavior] takes
+  /// effect. Min 5, Max 262116
+  ///
+  /// These overrides change when messages are sent. No need to mess with them
+  /// unless you're running into issues
+  int get maxBytesPerMessage => _maxBytesPerMessage;
+
+  set maxBytesPerMessage(int val) => _maxBytesPerMessage = boundValue(
+        val: val,
+        maxVal: awsMaxBytesPerMessage,
+        minVal: awsMinBytesPerMessage,
+      );
+
+  /// private maxBytesPerMessage
+  int _maxBytesPerMessage;
+
+  /// Changes how many bytes can be sent in each API request before a second
+  /// request is made. Min 1, Max 1048576
+  ///
+  /// These overrides change when messages are sent. No need to mess with them
+  /// unless you're running into issues
+  int get maxBytesPerRequest => _maxBytesPerRequest;
+
+  set maxBytesPerRequest(int val) => _maxBytesPerRequest = boundValue(
+        val: val,
+        maxVal: awsMaxBytesPerRequest,
+        minVal: awsMinBytesPerRequest,
+      );
+
+  /// private maxBytesPerRequest
+  int _maxBytesPerRequest;
+
+  /// Changes the maximum number of messages that can be sent in each API
+  /// request. Min 1, Max 10000
+  ///
+  /// These overrides change when messages are sent. No need to mess with them
+  /// unless you're running into issues
+  int get maxMessagesPerRequest => _maxMessagesPerRequest;
+
+  set maxMessagesPerRequest(int val) => _maxMessagesPerRequest = boundValue(
+        val: val,
+        maxVal: awsMaxMessagesPerRequest,
+        minVal: awsMinMessagesPerRequest,
+      );
+
+  /// private maxMessagesPerRequest
+  int _maxMessagesPerRequest;
+
   /// CloudWatchLogStack constructor
   CloudWatchLogStack({
     this.largeMessageBehavior = CloudWatchLargeMessages.truncate,
-  });
+    int maxBytesPerMessage = awsMaxBytesPerMessage,
+    int maxBytesPerRequest = awsMaxBytesPerRequest,
+    int maxMessagesPerRequest = awsMaxMessagesPerRequest,
+  })  : _maxBytesPerMessage = boundValue(
+          val: maxBytesPerMessage,
+          maxVal: awsMaxBytesPerMessage,
+          minVal: awsMinBytesPerMessage,
+        ),
+        _maxBytesPerRequest = boundValue(
+          val: maxBytesPerRequest,
+          maxVal: awsMaxBytesPerRequest,
+          minVal: awsMinBytesPerRequest,
+        ),
+        _maxMessagesPerRequest = boundValue(
+          val: maxMessagesPerRequest,
+          maxVal: awsMaxMessagesPerRequest,
+          minVal: awsMinMessagesPerRequest,
+        );
 
   /// The stack of logs that holds pre-split CloudWatchLogs
   List<CloudWatchLog> logStack = [];
@@ -30,7 +98,7 @@ class CloudWatchLogStack {
     for (final String msg in logStrings) {
       final List<int> bytes = utf8.encode(msg);
       // AWS hard limit on message size
-      if (bytes.length <= awsMaxByteMessageSize) {
+      if (bytes.length <= maxBytesPerMessage) {
         addToStack(time, bytes);
       } else {
         fixMessage(bytes, time, msg);
@@ -63,16 +131,16 @@ class CloudWatchLogStack {
         throw CloudWatchException(
           message:
               'Provided log message is too long. Individual message size limit is '
-              '$awsMaxByteMessageSize. log message: $msg',
+              '$maxBytesPerMessage. log message: $msg',
           stackTrace: StackTrace.current,
         );
     }
   }
 
   /// Truncates the middle of a message and replaces it with ...
-  static List<int> truncate(List<int> bytes) {
+  List<int> truncate(List<int> bytes) {
     // plus 3 to account for "..."
-    final double toRemove = (bytes.length + 3 - awsMaxByteMessageSize) / 2;
+    final double toRemove = (bytes.length + 3 - maxBytesPerMessage) / 2;
     final int toRemoveFront = toRemove.ceil();
     final int toRemoveBack =
         toRemove % 1 == 0 ? toRemove.ceil() : toRemove.floor();
@@ -84,12 +152,12 @@ class CloudWatchLogStack {
   }
 
   /// Splits message into smaller chunks
-  static List<List<int>> split(List<int> bytes) {
+  List<List<int>> split(List<int> bytes) {
     List<int> newBytes = bytes;
     final List<List<int>> res = [];
-    while (newBytes.length > awsMaxByteMessageSize) {
-      res.add(newBytes.sublist(0, awsMaxByteMessageSize));
-      newBytes = newBytes.sublist(awsMaxByteMessageSize);
+    while (newBytes.length > maxBytesPerMessage) {
+      res.add(newBytes.sublist(0, maxBytesPerMessage));
+      newBytes = newBytes.sublist(maxBytesPerMessage);
     }
     if (newBytes.isNotEmpty) {
       res.add(newBytes);
@@ -102,26 +170,19 @@ class CloudWatchLogStack {
   /// Adds a json object of [time] and decoded [bytes] to the last CloudWatchLog
   /// on the last [logStack] Creates a new CloudWatchLog as needed.
   void addToStack(int time, List<int> bytes) {
+    final Map<String, dynamic> log = {
+      'timestamp': time,
+      'message': utf8.decode(bytes),
+    };
+    // each message has 26 bytes overhead as per documentation
+    final int logSize = bytes.length + 26;
     // empty list / aws hard limits on batch sizes
     if (logStack.isEmpty ||
-        logStack.last.logs.length >= awsMaxMessagePerByte ||
-        logStack.last.messageSize + bytes.length > awsMaxByteBatchSize) {
-      logStack.add(
-        CloudWatchLog(
-          logs: [
-            {
-              'timestamp': time,
-              'message': utf8.decode(bytes),
-            },
-          ],
-          messageSize: bytes.length,
-        ),
-      );
+        logStack.last.logs.length >= maxMessagesPerRequest ||
+        logStack.last.messageSize + logSize > maxBytesPerRequest) {
+      logStack.add(CloudWatchLog(logs: [log], messageSize: logSize));
     } else {
-      logStack.last.addLog(
-        log: {'timestamp': time, 'message': utf8.decode(bytes)},
-        size: bytes.length,
-      );
+      logStack.last.addLog(log: log, size: logSize);
     }
   }
 
@@ -140,5 +201,14 @@ class CloudWatchLogStack {
   void prepend(CloudWatchLog messages) {
     // this is the fastest prepend until ~1700 items
     logStack = [messages, ...logStack];
+  }
+
+  /// Returns an int that has been bounded between [minVal] and [maxVal]
+  static int boundValue({
+    required int val,
+    required int minVal,
+    required int maxVal,
+  }) {
+    return min(maxVal, max(val, minVal));
   }
 }