+++ test/cloudwatch_handler_test.dart
@@ -1,4 +1,5 @@
 import 'package:aws_cloudwatch/aws_cloudwatch.dart';
+import 'package:aws_cloudwatch/src/logger.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -98,6 +99,36 @@ void main() {
       expect(cw.raiseFailedLookups, true);
       cw.raiseFailedLookups = false;
       expect(cw.raiseFailedLookups, false);
+    });
+    test('maxBytesPerMessage', () {
+      expect(cw.maxBytesPerMessage, awsMaxBytesPerMessage);
+      final CloudWatch instance = cw.createInstance(
+        logGroupName: 'logGroupName',
+        logStreamName: 'logStreamName',
+      );
+      cw.maxBytesPerMessage = 100;
+      expect(cw.maxBytesPerMessage, 100);
+      expect(instance.maxBytesPerMessage, 100);
+    });
+    test('maxBytesPerRequest', () {
+      expect(cw.maxBytesPerRequest, awsMaxBytesPerRequest);
+      final CloudWatch instance = cw.createInstance(
+        logGroupName: 'logGroupName',
+        logStreamName: 'logStreamName',
+      );
+      cw.maxBytesPerRequest = 100;
+      expect(cw.maxBytesPerRequest, 100);
+      expect(instance.maxBytesPerRequest, 100);
+    });
+    test('maxMessagesPerRequest', () {
+      expect(cw.maxMessagesPerRequest, awsMaxMessagesPerRequest);
+      final CloudWatch instance = cw.createInstance(
+        logGroupName: 'logGroupName',
+        logStreamName: 'logStreamName',
+      );
+      cw.maxMessagesPerRequest = 100;
+      expect(cw.maxMessagesPerRequest, 100);
+      expect(instance.maxMessagesPerRequest, 100);
     });
   });
   group('functions', () {